requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
alembic==1.12.0
black==25.1.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
greenlet==3.2.0
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
priority==2.0.0
PyJWT==2.8.0
pytest==7.4.0
pytest-asyncio==0.23.6
python-dotenv==1.0.0
PyYAML==6.0.2
Quart==0.20.0
quart-cors==0.8.0
Quart-JWT-Extended==0.1.0
sqlalchemy[asyncio]
ruff==0.11.6
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.20
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
wsproto==1.2.0
